+++ src/Testing/MakesJsonApiRequests.php
@@ -21,6 +21,7 @@ namespace CloudCreativity\LaravelJsonApi\Testing;
 use CloudCreativity\JsonApi\Testing\DocumentTester;
 use CloudCreativity\JsonApi\Testing\ErrorsTester;
 use CloudCreativity\LaravelJsonApi\Document\GeneratesLinks;
+use Illuminate\Contracts\Routing\UrlRoutable;
 use Illuminate\Http\Response;
 use Neomerx\JsonApi\Contracts\Document\DocumentInterface as Keys;
 use Neomerx\JsonApi\Contracts\Document\LinkInterface;
@@ -83,7 +84,7 @@ trait MakesJsonApiRequests
      * Assert response is a JSON API resource index response.
      *
      * @param string|string[] $resourceType
-     * @param string|int|null $resourceId
+     * @param string|int|UrlRoutable|null $resourceId
      *      if a singular resource is expected, the id of the singular resource.
      * @param string $contentType
      * @return $this
@@ -93,6 +94,10 @@ trait MakesJsonApiRequests
         $resourceId = null,
         $contentType = MediaTypeInterface::JSON_API_MEDIA_TYPE
     ) {
+        if ($resourceId instanceof UrlRoutable) {
+            $resourceId = $resourceId->getRouteKey();
+        }
+
         $this->assertJsonApiResponse(Response::HTTP_OK, $contentType);
 
         if (!$resourceId) {
@@ -229,6 +234,10 @@ trait MakesJsonApiRequests
      */
     protected function seeStatusCode($expected)
     {
+        if (!$this->response) {
+            PHPUnit::fail('No response - have you made a call to the application?');
+        }
+
         $actual = $this->response->getStatusCode();
         $message = "Expected status code {$expected}, got {$actual}";
         $content = (array) json_decode((string) $this->response->getContent(), true);